example-instrumented-app/requirements.txt
opentelemetry-api>=1.13.0
opentelemetry-exporter-otlp-proto-grpc>=1.13.0
opentelemetry-sdk>=1.13.0
